Make the sidebar behave like a real one, plus JS refactor
