+++ assets/js/components/Layout/index.js
@@ -1,0 +1,3 @@
+import Layout from './Layout';
+
+export default Layout;
